Store Sales and Profit Analysis using Python
Store sales and profit analysis is the task of analyzing the performance of a retail store in terms of its sales and profits. It helps businesses identify areas for improvement and make data-driven decisions to optimize their operations, pricing, marketing, and inventory management strategies to drive revenue and growth. So, if you want to learn how to analyze the sales and profit of a store, this article is for you. In this article, I will take you through the task of Store sales and profit analysis using Python.
Store Sales and Profit Analysis: Dataset
For the task of Sales and Profit Analysis, we need to have a dataset with time-period data, sales data, product data, pricing data, customer data, etc.

I found an ideal dataset for this task on Kaggle. You can download the dataset from here.

In the section below, I’ll take you through the task of Store Sales and Profit analysis using Python.